PCVK/requirements.txt
gradio>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-image>=0.21.0
scipy>=1.11.0
rembg>=2.0.50
tqdm>=4.65.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv
onnxruntime
matplotlib
tensorboard

# Azure dependencies
azure-storage-blob
azure-identity

# Firebase dependencies
firebase-admin